+++ packages/ethers/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "ethers/5.5.2";
+export declare const version = "ethers/5.5.3";
 //# sourceMappingURL=_version.d.ts.map