+++ packages/providers/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "providers/5.5.1";
+export declare const version = "providers/5.5.2";
 //# sourceMappingURL=_version.d.ts.map